Introduce ptr_provenance for LoadInst/StoreInst
The ptr_provenance tracks dependencies on noalias intrinsics without
blocking optimizations on the pointer value.

Note: this implementation might have some speed effect on all instruction
because of the extra bit that is used to track if the ptr_provenance is
present or not. This might impact the computation of the number of arguments.

[PtrProvenance] Helpers for accessing and merging PtrProvenance

FIXUP: llvm/include/llvm/IR/Metadata.h

[Instructions] copyOptionalPtrProvenance

[PtrProvenance] Use UnknownProvenance when cloning a load/store instruction with provenance